Calculate dumpsite scores

Calculate dumpsite scores

Usage

assess_dumpsites(
  df_point,
  df_reach,
  start_date,
  end_date,
  reach_prefix_from_table,
  reach_prefix_from_layer,
  point_prefix_from_table,
  point_prefix_from_layer
)

Arguments

df_point

Point data from Rapid Stream Assessment (RSA) dataset.

df_reach

Reach data from Rapid Stream Assessment (RSA) dataset.

start_date

Beginning of timeframe being evaluated (ex. first day of five-year assessment period).

end_date

End of timeframe being evaluated (ex. last day of five-year assessment period).

reach_prefix_from_table

Field name prefix sourced from RSA reach attribute table. Should be consistent with the name of the stream reach attribute table from the RSA geodatabase.

reach_prefix_from_layer

Field name prefix sourced from RSA stream reach layer. Should be consistent with the name of the stream reach polyline layer from the RSA geodatabase.

point_prefix_from_table

Field name prefix sourced from RSA point attribute table. Should be consistent with the name of the point assessment attribute table from the RSA geodatabase.

point_prefix_from_layer

Field name prefix sourced from RSA point layer. Should be consistent with the name of the intersected stream point layer from the RSA geodatabase.

Value

List of 2 dataframes: summary and score.

Examples

assess_dumpsites(df_point_example,
                           df_reach_example,
                           "2020-01-01",
                           "2024-12-31",
                           "StreamReachAttributes",
                           "StreamReaches",
                           "StreamPointAttributes",
                           "StreamPoints_Intersect" )
#> $summary
#> # A tibble: 7 × 8
#> # Rowwise: 
#>   sci_subshed            dumpsite_count applied_influence subshed_length_meters
#>   <chr>                           <int>             <dbl>                 <dbl>
#> 1 Broad Branch                        1                -1                 9090.
#> 2 Fort Davis Tributary                1                -1                 1198.
#> 3 Fort Dupont Tributary               1                -2                 8637.
#> 4 Fort Stanton Tributary              1                -2                 2214.
#> 5 Oxon Run                            1                -1                 9965.
#> 6 Rock Creek                          2                -2                32011.
#> 7 Watts Branch                        1                -2                 6013.
#> # ℹ 4 more variables: subshed_length_miles <dbl>, sites_per_mile <dbl>,
#> #   score <dbl>, score_weighted <dbl>
#> 
#> $score
#> # A tibble: 25 × 2
#> # Rowwise: 
#>    sci_subshed            Dumpsites
#>    <chr>                      <dbl>
#>  1 Broad Branch                   8
#>  2 Fort Davis Tributary           2
#>  3 Fort Dupont Tributary          7
#>  4 Fort Stanton Tributary         4
#>  5 Oxon Run                       8
#>  6 Rock Creek                     7
#>  7 Watts Branch                   6
#>  8 Battery Kemble Creek          10
#>  9 Dalecarlia Tributary          10
#> 10 Dumbarton Oaks                10
#> # ℹ 15 more rows
#>